+++ src/modules/forms/actions.js
@@ -9,7 +9,7 @@ function reset(name) {
   };
 }
 
-function requestForms(name, page) {
+function requestForms(name, page, query) {
   return {
     type: types.FORMS_REQUEST,
     name,
@@ -33,10 +33,10 @@ function failForms(name, err) {
   };
 }
 
-export const indexForms = (name, page = 1, params = {}, options) => {
+export const indexForms = (name, page = 1, params = {}) => {
   return (dispatch, getState) => {
     dispatch(requestForms(name, page));
-    const forms = selectRoot('forms', getState());
+    const forms = selectRoot(name, getState());
 
     if (parseInt(forms.limit) !== 10) {
       params.limit = forms.limit;
@@ -45,7 +45,10 @@ export const indexForms = (name, page = 1, params = {}, options) => {
       params.skip = ((parseInt(page) - 1) * parseInt(forms.limit));
       params.limit = parseInt(forms.limit);
     }
-    const formio = new Formiojs(options.project + '/form');
+    else {
+      delete params.skip;
+    }
+    const formio = new Formiojs(Formiojs.getProjectUrl() + '/form');
 
     return formio.loadForms({params})
       .then((result) => {