+++ Changelog.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 3.1.3
+### Added
+ - Url to reducers
+
+### Changed
+ - isFetching becomes isActive
+ - FormEdit will autogenerate name and path for new forms.
+
+### Removed
+ - Options parameter to actions.
+
 ## 3.1.2
 ### Added
  - New reset actions for resetting state